Reviewed-on: http://git.axibug.com/sin365/AxibugEmuOnline/pulls/21

+++ AxibugEmuOnline.Client/Assets/VirtualNes.Core/Mapper/Mapper023.cs
@@ -28,7 +28,7 @@ namespace VirtualNes.Core
 		{
 			addrmask = 0xFFFF;
 
-			for (INT i = 0; i < 8; i++)
+			for (byte i = 0; i < 8; i++)
 			{
 				reg[i] = i;
 			}
